Uploaded Project Files
These are the project files used for the Master Controller of the project. The project was managed manually without a source control system and logs were kept in "GROUP 2 README.docx" file. These set of files were the final iteration (version) of the project before demonstration. An additional pdf file is provided which combines all the code from all the files in one document.